Rename `FewShot` to `MultipleChecker`
The name `FewShot` doesn't really make sense any more, as the class
isn't using a few-shot prompt any longer.

Really what this class is doing is taking an input string and the name
of a config file and asking the LLM to check the input against multiple
rules in that config file.

If there are no matches, it returns the string `False | None`.

If there are matches, it returns the string `True | "1, 2, 3"`, i.e. a
numeric list of the rules that were matched.

So it's basically checking the input against multiple rules, hence the
name `MultipleChecker`.