Add Answer#use_in_rephrasing? and Question#use_in_rephrasing?
We don't want to pass answers that have been aborted due to forbidden words,
question routing or guardrails to the rephraser.

This adds a method to answer which filters out answers with the following
statuses:

```
abort_answer_guardrails
abort_forbidden_terms
abort_jailbreak_guardrails
abort_output_guardrails
abort_question_routing
```

It also updates the Question model to delegate to the answer model when
the 'use_in_rephrasing?' is called on a question object.