Refactor question router handling for using an answer
This principally makes use of a `use_llm_answer?` to determine whether
we use the answer response from the LLM call. It did morph into a bit
more of a refactor as I realised there was a risk of an error if we got
a long answer back with a genuine_rag label and needed to handle that as
the first conditional otherwise we risked [1].

I had to do a bit more work on the tests as some scenarios were missing
(genuine_rag and finish_reason == "length") - which I understand were
lost in a rebase.

[1]: https://github.com/alphagov/govuk-chat/blob/985461912467c2d0e3c59726fdb0d6f639088763/app/models/answer.rb#L32