Add .aggregate_export_data to EarlyAccessUser
We're going to start passing some aggregate data to BigQuery for user
counts.

This commit adds a .aggregate_export_data method to the EarlyAccessUser
mode which will be used by the BigQuery::Exporter to export the data.

This collates early access user and deleted early access user data into a
hash.