Define answer class in data migrations that update statuses
When these migrations run, the correspoding statuses have already been
removed from the enum defined in the answer model and blow up.

This adds an Answer class within the migration to ensure that if the
application codes enum is out of sync with the database, we still
have access to the functionality provided by the Rails enum.