Add WaitingListUser.aggregate_export_data
This follows the established pattern of using a class method to aggregate
waiting list user data which will be used to export to BigQuery.

We don't want to pass the number of waitilist list users promoted to
BigQuery since that's captured in the early access user stats so i've
ensured that's filtered out.