Extract uploading file to BigQuery into a separate class
The Bigquery::Exporter was doing two main things:

1. collating data to be uploaded to bigquery into tmp files
2. configuring and uploading the files to bigquery

This breaks the the second responsibility into a a Bigquery::Uploader
class.

This class takes a table_id and a tmpfile. It then:

1. creates the table if one isn't already present
2. configures the load_job
3. uploads the file to bigquery
4. raises an error if the upload fails

This leaves the exporter class to do the following for each model:

1. grab the data required
2. strip nil values if it's a top level model
3. create a tmp file with the table data converted to JSON
4. call the uploader
5. create a new BigqueryExport record
6. return an object with details of the files uploaded

That's still quite a bit of responsibility, but it's a lot better than
previously.