Rename `question_limit` to `individual_question_limit`
This better represents the use of this field as it doesn't represent the
users actual question limit and only when it is applied individually.
Changing the name of this field lets us create a method of
question_limit on a user.

This rename will risk 500 errors for the users that are using the app
while this change takes place - however as we're not yet live I'm going
to go ahead anyway.