Add a too_many_requests response
This adds a view and an errors controller method for a 429 too many
requests response.

I'm expecting to use this as part of configuring Rack::Attack, either by
rendering the view directly from Rack::Attack or raising an error to
trigger a Rails rescue response.

I've used copy for this from static [1]

[1]: https://github.com/alphagov/static/blob/94a64738f14f4558cde9d3e2998dc8543f047f61/app/views/root/429.html.erb