Configure Faraday middleware to parse JSON and raise errors
This follows guidance in the following docs

https://lostisland.github.io/faraday/#/middleware/included/raising-errors

which outlines how to raise errors when a response is not successful.

And https://lostisland.github.io/faraday/#/middleware/included/json
which outlines how to parse JSON in the body of responses by default.

I've that the JSON is parsed successfully via manually hitting the API
via console and it's working as expected.