Reorder the BigQuery exporter
We want to order the export so that the things most likely to fail
at the start so that:

- it fails fast if there is a problem
- we don't push data for some tables to BigQuery and not for others

This addresses the first point. A follow up commit will ensure that
we store the tempfiles, then iterate over them and push them once we
have all the required files for each table.