Add deleted_by_admin_user_id field to DeletedWaitingListUser and DeletedEarlyAccessUser
Adds a nullable field to both DeletedEarlyAccessUser and DeletedWaitingListUser to store which admin
has deleted them in the scenario that they are deleted by an admin (both can be done by the user
themselves by unsubscribing).

The references to admin_user_id are not foreign keys, as we don’t want these records to block
deleting admin users.