Don't create WaitingListUser if waiting list is full
If the waiting list is full and no instant access places are availble,
we don't want to create the users account (early access or waiting list).

Instead we want to render a page letting the user know there are no places
available.

This updates the reason for visit form and update endpoint to handle this.