Setup magic link limit to rate limit spam
This adds a limit to the number of magic links that can be sent to a user
a 5 minute period. This is to prevent someone spamming users with magic links.

The sign_in_or_up form now checks if 3 magic links have been sent in the
last 5 minutes. If so, we return a result object with the :magic_link_limit_reached
outcome.