Use a config file to manage the user research questions
This is to provide a centralised place for them to be managed and avoid
duplication.

I wasn't 100% sure I could use data for Rails.configuration as part of
class methods that are eager loaded but makes sense that config is
loaded first, since that is where eager loading is configured.

I've made use of fetch in the specs to raise an error if a config value
is nil.