Update pipeline to include metric for cached tokens
We want to keep an eye on the amount of tokens we're caching since
they're 50% discounted. This commit adds a metric for all OpenAI calls to
track it.

From my testing, it looks like the jailbreak and rephraser consistently
return 0. But i think it's worth adding them anyway as we may well tweak
the prompts down the line which may cause caching to take effect.