Extract the logic signing up user into it's own model
This is a preliminary bit of work for adding an additional UR question so
the signup flow. I've added it as a separate PR to reduce the size of the
PR which will add the new questions endpoints.

At the moment signing up a user is coupled to the reason for visit
form/endpoint which is an issue as we want to add a third question.
Rather than moving the logic in to the new questions form, it makes more
 sense to extract it into it's own model.

If we don't do this it means each time we add or remove additional questions
we will have to move the logic for signing (or not signing up) a user
to the last questions form.

By moving it to it's own model we can simply

- add/remove the questions form which validates the users input
- update the PilotSignUp model to handle the additional question value
- move the logic in the controller that handles the result of a user signing up the last questions POST endpoint