Add additional UR question endpoints
Apologies this is quite a noisy commit due to all the test updates
required (they're pretty much a copy paste job). I've added both the
found_chat endpoints in this commit so the:

- routes
- controller actions
- views
- form

etc.

I've also extended the PilotSignUp model to include persisting the
new questions value to the user when created.

Finally i've updated the tests to reflect the new flow.