Extend ensure_sign_up_flow_position to handle found_chat
This updates the ensure_sign_up_flow_position method to handle a user arriving
at the found_chat endpoints without a user description or reason_for_visit
value set in the sign_up session.

If a user description is not found it redirects to the first question and
if a user description is present and a reason_for_visit is not present it
redirects to the second question.

I've moved the 'render_not_accepting_signups_if_sign_ups_disabled' before
action above the 'ensure_sign_up_flow_position' one to make sure it only
runs when pages are being rendered and not seeing the sign up disabled
page.