Refactor PilotSignUp class
This simpiifies the PilotSignUp class. It:

- removes the ActiveModel::Model and ActiveModel::Attributes modules and
instead passes the attributes to the constructor
- extracts logic for creaitng for checking if the waiting list is full,
and creating early access and waiting list user into private methods
- updates the result object to take a sesion so that we don't need to use
ivars to capture the session created within the lock block
- updates the tests to use .call instead of .new and #call

It also updates the SignUpController to pass in the session sign up variables
as kwargs to the PilotSignUp class.