Sanitise user input to prevent ASCII smuggling in questions
This commit runs sanitisation prior to validations in `Form::CreateQuestion`.

Sanitisation removes characters in the range U+E0001 to U+E007E, which are
unicode characters originally intended for tagging texts by language. These
invisible characters allow "ASCII smuggling" attacks on LLMs where they can
be hidden inside legitimate input and so offer an alternative route
for prompt injection.

In cases where characters in this range are removed by sanitisation, the
original message will be saved in the `unsanitised_message` field on the
answer model, with the smuggled characters converted to visible unicode.