Allow nil input to user_research_question_option_text
This helper didn't anticipate that it can easily be called with a nil
argument for an option value, as neither of the models that make use of
it require a non-nil value.

Rather than adding conditionals in the view logic that calls this
function it seemed more elegant to have this case handled by the helper
directly.