Configure Rack::Attack to raise Rails' rescued exceptions
This configures Rack::Attack to raise a ThrottledRequest exception when
it throttles a request, this exception is configured to be a rescue
response so it will render a Rails error response.

Using this approach saves us having to configure a means to render a
view as part of the Rack::Attack code.