Throttle homepage post requests to 10 in 5 minutes
This is a measure to limit the amount of requests a single users IP can
make to the homepage endpoint which acts a sign-in or sign-up entry
point.

The reason we want to limit requests to this endpoint is that it could
be used to try to brute force enter the application by lots of requests
for known email addresses or it could be used to send unwanted emails to
users with accounts - by creating them new magic links.

Setting what the limit should be is rather delicate as there is the
possibility of organisations/ISPs sharing IP addresses across multiple
users. We've gone for a limit of 10 requests in 5 mins on the
expectation that this should be more than ample for an individual user
yet still suffice for most cases of a shared IP address.

In testing these I have gone with the rather basic assertion that a
request is either returning 429 or any other status code. I chose to go
down this rather simplistic avenue after hitting pitfalls of trying to
ensure a response was a success - particularly after considering the
other endpoints we intended to throttle. I found that it required
applying the correct parameters and, for non idempotent endpoints,
rotating the parameters on each request. I decided it was simpler to
take this route knowing that it did have some risks that it didn't
exercise the endpoint being operational.

For testing these we need to an active cache store available. This is
because Rack::Attack uses the cache store as a means to track requests
across distinct application servers. As we test with a null cache store
[1], I have setup a RSpec flag to apply an active cache for the duration
of the test.

[1]: https://github.com/alphagov/govuk-chat/blob/494be4bde584176cefb23f9dcfac50055e81f09c/config/environments/test.rb#L26