Throttle by IP endpoints with token authentication
This applies rate limiting by IP address to the endpoints we have that
use a token to identify a particular user. The motivation for this is to
limit the ability for an individual IP address to try brute force any of
these identifiers.

I've used a more lenient limit than we have on the other limits as these
endpoints only have side effects if a correct value is used and thus we
can reduce the risk that a shared IP could be impacted.

It was more challenging to use the Rails routing system to match
requests than it was for post requests as the path contains parameters.
I used an approach from the Rack::Attack documentation to match a Rails
route by URL [1]. In my testing there were a couple of situations where
we get exceptions 1) route not matching (ActionController::RoutingError)
and 2) errors related to constraints of matching (in particular I was
receiving an error because warden didn't exist for [2]). As it seems
hard to predict if we'll get constraint related errors I've gone for the
catch of StandardError.

[1]: https://github.com/rack/rack-attack/blob/main/docs/advanced_configuration.md#match-actions-in-rails
[2]: https://github.com/alphagov/gds-sso/blob/cae5114a3f01e91fb4afb8b3704e964c5c3eb1da/lib/gds-sso/authorised_user_constraint.rb#L9-L10