Update from params to pass empty hash if not present
When you submit a form with an unselected radio/checkbox no value is
passed in the params.

This causes the params.require(:form) to blow up. To get round this
we were submitting a hidden field with a default value of nil for the
choice.

This was failing silently for the last question due to a typo and we
hadn't realised.

To get round this we can use fetch to check if the form is present in the
params and if not pass an empty hash.

I've updated one of the tests to reflect the actual behaviour.