Remove redundant assertion in waiting_list_users_spec
This is a hangover from updating this system to remove adding a reason for
visit. But i forgot to remove the assertion that was checking for it's
presence. It's still passing because the assertion is checking for nil
in the content, but we're getting the following stdout

```
Checking for expected text of nil is confusing and/or pointless since it
will always match. Please specify a string or regexp instead.
/govuk/govuk-chat/spec/system/admin/waiting_list_users_spec.rb:53
```

I've also removed the selecting how they found chat to make it clear that
these were skipped on purpose.