Update ur question validations for inclusion in enum keys
This updates the validations for the UR questions radios to check for
inclusion in the corresponding enums keys rather than just checking
for presence.

This should really only get triggered for the presence validation since
we use the config to generate the radios. And this config is tested against
the enum keys.

So for that reason the message rendered should remain the same since it's
been provdided by the CDs.