Use Client-True-IP header for rate limiting via IP address
This resolves an issue identified in the testing of rate limiting that
the IP address that is used as a discriminator was the Fastly IP address
rather than the end users' IP.

The reason the wrong IP occurs is based on the data proxied through
nginx for a request from me in the WCB building the headers passed are:

Remote-Addr: 10.1.20.23
X-Forwarded-For: 217.196.229.79, 157.52.69.51

Where 157.52.69.51 is the Fastly IP.

Looking at the Rake::Request IP logic [1] we can see how this results in
the Fastly IP being the one picked (as Remote-Addr is an ignored proxy)
and all of X-Forwarded-For are trusted, thus last one used).

To resolve this we need a different way to reliably get the client IP.
The way that seemed most appropriate was to use the header set by the
CDN at the boundary [2].

I've not set any fallback ips (such as request.ip) as I thought it was
safer that, in the absence of the CDN header, to not fallback as it
increased the chance of a false positive blocking.

It's worth noting that we should not use this header in scenarios
without a CDN in front of us that sets this header. This is because this
is not a header typically set by a client, and without anything
controlling it at a boundary a user could just change it.

[1]: https://github.com/rack/rack/blob/fc541f3f443ed170b9b1daed4d16c6407a12573d/lib/rack/request.rb#L414-L433
[2]: https://github.com/alphagov/govuk-fastly/blob/7aac937c9ed5ae036505e2f8a484629389fdcf37/shared/_boundary_headers.vcl.tftpl#L8