Add a last 24 hours view to admin metrics
This modifies the metrics endpoint to offer views for last 24 hours in
addition to last 7 days.

It adds the secondary navigation component to provide the choice between
the two time periods.

For a few data groups with very specific data (abort/error statuses,
guardrails failed, question labelling) I felt that these would be
unusable as bar charts over a 24 hour period so configured them to be
pie charts that no longer show a grouping by hour. I have also modified
the rendering of these to be done in 2 in a row columns rather than full
width as they are for 7 days - which required some delicate view
editing.

Apologies that is a bit of a monster commit.