Replaces revoked at column with struck through users with labels
This removes the revoked at column from the table on the early access
users table. It is replaced by the email address being struck through
when a user is revoked or has been shadow banned.

If they've been shadow banned, i've appened (shadow banned) to their
email address. If they've been revoked I've appened (revoked).

There was quite a bit of logic building up in the view for rendering
the email and questions column so i've moved that into a helper methods.