Reorder admin settings
This orders the settings to be more logically connected. This has both
the significant enabling settings up top, instant access on it's own
ending with the three related to waiting list at the bottom.

The tests have been re-ordered to follow the same flow as the view. This
has no material impact on the application, it's just done to feel
logical to developers. Previously they were ordered based on which
setting was added first.