Add restore_confirm endpoint
This commit changes the restore endpoint to be a get rather than patch
and adds a view. This view explains what the restore action does and
provides a textarea for the admin to enter a reason for restoring the
user.

It then adds a PATCH restore_confirm endpoint that takes the reason
entered by the admin and passes it with the user to the restore form.