Make purpose of settings in request tests clearer
The settings model defaults to being created with sign_up_enabled set to
false. Thus for tests related to sign up we need to ensure a different
model is created via a factory.

Previously this was a bit cryptic as it was done in a before block for
an individual request test and meant using a context in a different
scenario would fail without it. This moves the code into the context
directly with an explicit sign_up_enabled attribute to indicate purpose.

A potential risk with this approach was if another test also runs
`create(:settings)`. However I checked this and this being run twice
won't generate an error and new attributes will be modified.