Shadow ban user when jailbreak threshold is reached
This commit does 2 things:

- Adds a handle_jailbreak_attempt method to the early access user

This incrments the 'bannable_action_count'. If the count reaches the threshold
defined in the BANNABLE_ACTION_COUNT_THRESHOLD constant then it updates
the shadow_banned_at and shadow_banned_reason attributes.

In future the slack message will also be added to this method.

- Updates the ComposeAnswerJob to call the handle_jailbreak_attempt method
if a user is present for the conversation and the answer has the
"abort_jailbreak_guardrails" status.