Add system spec for a user asking a question while shadow banned
This adds a pretty basic system spec that covers the happy path of a user
being shadow banned and asking a question.

We could add a more involved spec if we want that covers a 10th jailbreak
attempt occurring and that the subsequent question makes no external calls,
but i think this is probably fine as we've got that unit tested elsewhere.