Use groupdate to populate metrics data without values
This replaces the custom code I wrote to populate data periods with a
count of zero. I had not realised that the group_by_* methods of
groupdate could accept values of `last` or `range` [1] which meant I
didn't need to write the custom code. It also has a format argument
which could manage the timestamp formatting [2].

There were a couple of quirks to sort out when applying this. The first
was that groupdate is inconsistent with whether it includes data or not
depending on the amount of groups there are. I have added a method to
make it always be empty for no data. The second was that I wasn't sure
of a shorthand to set the time to a formatted string so used a lambda.

The motivation for this change was to investigate moving the custom code
to a model for easier direct testing as it is hard to understand [3].

[1]: https://github.com/ankane/groupdate?tab=readme-ov-file#time-range
[2]: https://github.com/ankane/groupdate?tab=readme-ov-file#keys
[3]: https://github.com/alphagov/govuk-chat/pull/864#discussion_r1827490103