Use escaped_simple_format for access reasons on show page
At the moment, we don't preserve newlines in the revoked_reason,
shadow_banned_reason or restored_reason values on the early access user
show page.

This updates them to use the escaped_simple_format helper to preserve them.

I've had to override the margin top to 0 to fix issues, but it's working
nicely. The p tags still preserve bottom margin to it's clear where the
linebreaks are.