Configure instant & delayed access places schedule for live
Now that we have a go live date (7th November) and an updated user
volume plan [1] we can configure these schedule constraints according
to the plan.

For instant access users, we expect to start allocating places from 7th
November and to not allocate places after December 1st. The highest
number of places in the volume plan is 150, so I've set 200 as the max
to offer some room.

Delayed access places will be allocated starting from Monday 11th
November and will also finish on the 1st December. We expect to
allocate these only once a day so have higher figures here. The maximum
figure in the volume plan is 600, so I set 1000 as the max to offer some
room.

[1]: https://docs.google.com/document/d/1ero2mqJ6FlWNV4uz6Hxof5x5cyPkmKYefEpWdkHt5JY/edit?tab=t.0