Add method to notify Slack when a user is shadow banned
Creates a method which uses the slack-poster gem to notify slack
with a simple message about a user being shadow banned.

The message links to the user's profile in the admin section.

It does nothing if the `AI_SLACK_CHANNEL_WEBHOOK_URL` env var is blank.
This means we can call the method in all environments but it will only
actually run in production where the env var is set.