Fix flaky search result spec
This spec asserts the URL that an OpenSearch search result has in the
admin panel. That URL contains a param of `score_calculation` which is a
URL-escaped string of `score * document_weight_type = weighted_score`.

This spec has been failing on CI because it relies on the fact that the
search result has a score of 0.9 (with a few more decimal places) but
recently it's been coming back with a score closer to 1.

We probably shouldn't be relying on the score always being the same. So
instead we can get the actual value by performing a search within the
spec and grabbing the score from that, then using that score to
calculate the weighted score too.