Add a focusable only class for 'Start new chat' link
This adds a class that will make the 'Start new chat' link be invisible
unless it is focused. This is set when a user has a conversation that
does not yet have any messages so that they can't see the button that
isn't beneficial to click (it'll just redirect you back to the same
page).

This class is intended to be toggled by JavaScript at the point of when
a user changes from having a non active conversation to an active one.

The reason this has been added as an invisible element rather than an
element that is just added dynamically is for users on screen readers.
These users would not have the adding of this link announced to them so
would not know it is there.

Instead we have decided that we will always have the link there, so it
is announced to screen reader users, but is hidden for users who aren't
using a screen readers. With it being a link that is announced we need a
user to be able to reach the link with a tab, hence why it becomes
visible if it is focused.

Co-authored-by: James Chan <james.chan@digital.cabinet-office.gov.uk>