Make 'Start new chart' link visible by JS
This uses an event listener on the document element as a means to
communicate an event between two distinct modules so that they don't
have a direct coupling.

It uses the acceptance of a question as an indicator that a conversation
is active. This event will be triggered every time a question is
accepted and not just the first time. Should we have a need for only the
first question we'll have to add a mechanism to determine it.

This code is based on an initial implementation by Jon in PR 882 [1]

[1]: https://github.com/alphagov/govuk-chat/pull/882

Co-authored-by: Jon Kirwan <jon.kirwan@digital.cabinet-office.gov.uk>