Reduce mocking in guardrails rake spec
We had an issue recently where an additional mandatory argument was
added to the `Guardrails::MultipleChecker.call` method.

The `guardrails.rake` file wasn't updated to include this new argument
though, but the tests still passed so it wasn't picked up until somebody
tried to actually run the rake task.

The problem is the amount of mocking that's happening in the spec,
specifically the line which mocks the actual guardrail checking:

```
allow(Guardrails::MultipleChecker).to receive(:call).and_return(false_response, true_response)
```

This commit does a bit of refactoring to reduce the amount of mocking so
we hopefully won't see this problem again.

Really all we care about testing here is that the rake task performs the
evaluation and either prints the results to stdout or to a file. The
actual guardrailing logic is well tested elsewhere, in both the
`Guardrails::Evaluation.call` method and the
`Guardrails::MultipleChecker.call`.

So we don't really need to make the test produce a mix of true and false
responses. We just want to assert that:

* the right number of results are produced based on the number of lines
  in the input CSV file
* the results are merged with the model name,
  `average_prompt_token_count` and `max_prompt_token_count`
* the results contain the output string of the guardrailing (e.g. `True
  | "1"`)

So to make it simple, this commit stubs the guardrailing at the OpenAI
level so that the actual application code is tested, then verifies the
output as it did before.