Add waiting_list_full Slack notification
Adds a new notification to alert us when the waiting list is full.

This commit refactors the `SlackPoster` module to be a class instead,
which makes for a lot less code duplication.

For both notifications we want to include logic to do nothing if the
webhook env var isn't set. So making this a class allows this to be
added in a nicer way.

It also removes the `test_mode` logic. This was in place to make sure
the Slack integration was working so we don't really need to include it
in all of the message-sending methods.

Instead there's a new `.test_message` method which we can use in future
if needs be.