Merge guardrail config files
These were extremely similar with only minor differences. All of the
guardrails with the exception of sensitive_financial_matters are currently either:

1. only used by 1 type of guardrail (answer or question routing)
2. used by both but are identical

This commit merges the files by:

- having shared guardrails config defined
- having answer guardrails and question routing guardrails inherit/alias
the shared guardrails config where required
- adding unique guardrails to the answer guardrails and question routing
config

One thing to note i've gotten rid of the concept of guardrail mappings.
Now we have a list of guardrails that are always used defined, we can
just iterate through the away to get the guardrail number dynamically.