Update multiple_checker to use new config format
This updates the checker handle config use the new combined config.

The main changes are that the guardrails are now stored in an array.
This is uses to dynmically create the guardrails for the prompt by:

1. Grabbing the guardrails from the config
2. mapping across them and creating the content for the prompt for each
guardrail by:
  a. looking up the guardrail definition in the config
  b. prepending the index + 1 of the guardrail to the content

When the #extract_guardrails method is called, it now infers which
guardrail has been triggered by the index of the guardrail in the
the config + 1 so it matches the index of the guardrail in the prompt.