Replace tokenised link text with "source"
For context: we include a bunch of search results in the prompt we send to
the LLM to help it answer the question. Each result has a corresponding
URL which we show to the user so they can find out more information
about it.

We encountered some issues where if we sent the actual URL to the LLM,
sometimes it would rewrite it and thus return a link to a page that
doesn't exist. So to remedy that, we replace all URLs with tokens before
sending them to the LLM (e.g. "link_1") then rewrite them when we get the
response back.

Most of the time, the LLM is good at returning markdown links with some
relevant text as the link text. However sometimes it returns a link like
so:

```md
Most businesses register as a sole trader, limited company, or partnership. ([link_1][1]).
```

When we convert this into HTML, the user sees

`<a href="{url}">link_1</a>`

which looks like something is broken.

This commit adds a bit of logic when we're rewriting the href attributes
of a URL. We inspect the children (i.e. the link text nodes) and if any
of them match the token format (i.e. "link_\d+") then we rewrite that
link text to read "source" instead. This means the user sees this
instead:

`<a href="{url}">source</a>`