Capture additional search duration metrics
We want to capture some of the timings of sub components/interactions in
the search pipeline. This commit:

1. adds a 'metrics' attribute to the ResultSet data class
2. populates the metrics attribute with the embedding_duration,
search_duration and reranking_duration to the ResultSet instance returned
by the the ResultsForQuestion class
3. adds the metrics to the ResultSet instance that is returned by
ResultsForQuestion.call
4. updates the search result fetcher pipeline step to merge the additional
metrics when into the hash constructed in the build_metrics method.

This means that the metrics passed to the assign_metrics method are

{
  duration: float,
  embedding_duration: float,
  search_duration: float,
  reranking_duration: float
}