Rename answer_composition class to clock
Now we're using the monotonic_time method outside of the
answer_composition directory it doesn't really make sense for this
to be in a module called answer_composition.

Particularly because the ResultsForQuestion class is also used in the
admin UI search page.

This renames it to Clock to decouple it from answer composition.