Set answer sources as unused in forbidden terms checker
We've seen an example on production where a user asked a question which
got answered, but that went on to trigger the forbidden words check.

That in itself is fine, but the answer contained links to some
sources that the LLM used to generate the answer. The user was presented
with a canned response from the ForbiddenWordsChecker, but also a list
of the sources used to answer the question.

We don't want show any sources cited if we're returning this canned
response. So we need to set the sources to `used: false`.