Send feedback request email when user first asks a question
When a user has asked their first question, we enqueue an email to be
sent 10 minutes in the future asking them for feedback.

To test the 10 minute delay, I've set the `queue_adapter` to `:test` in
the spec helper file. This gives us access to nice helper methods like
`have_enqueued_mail`, `have_enqueued_job` and so on.

This commit also configured RSpec to clear the Sidekiq queue before each
spec. I had some issues with some specs where a job to send an feedback
request email was put on the queue in a test, then later on a system
test was run which calls the `execute_queued_sidekiq_jobs` method. So it
would try to process a job from a different spec and fail because the
record didn't exist in the database any more.