Configure test queue adapter for specs
This sets the `ActiveJob::Base.queue_adapter` to `:test` for all specs
that are _not_ system or request specs.

This means that jobs enqueued within these tests won't be put on the
Sidekiq queue, but on an internal memory queue instead.

Setting this value allows us to use the more concise RSpec methods of
`enqueue_job`, etc to assert that jobs have been enqueued, rather than
having to dig into the internals of Sidekiq to pull the job out.

We don't want to set this on system/request specs as we rely on jobs
being enqueued with Sidekiq and then executed in those specs.