Update get started email content to mention feedback
We're trying to drive more engagement with the survey. This updates the
get started email to mention the survey and ask users to provide feedback
when they've finished their convo.

Content here
https://docs.google.com/document/d/1pqZuqP0WzlwRU1-NDM3n0f8rLOcFpfkxRRTPfTFrpkw/edit?tab=t.0#heading=h.8766heg9q6rf

I've made a one word change to upate

```
Use the 'Share your feedback' button in the footer
```

to

```
Use the 'Share your feedback' link in the footer
```

since it's a link not a button.