Single constant to list Bigquery export information
This adds a new constant, TABLES_TO_EXPORT, which stores a list of all
the tables that are set-up in BigQuery. This will soon replace the two
constants above it.

This also changes the name of tables that store aggregate statistics to
reference that in their names. This has been done so that we can free up
the table name so they can be used for exporting data (as we intend to
with early_access_users).

The motivation for this was that we are going to need this information
in more places with the refactoring we're working on and that we noticed
a bug where "smart_survey_responses" wasn't included in resetting.

Co-authored-by: Chae Cramb <chae.cramb@digital.cabinet-office.gov.uk>