Bigquery::IndividualExport class
This creates a class to perform the individual exporting of a table into
data that can be imported into Big Query.

The motivation for this is to make this code reusable so it can be used
as both part of an hourly export and also be used to backfill data that
was missing.

Co-authored-by: Chae Cramb <chae.cramb@digital.cabinet-office.gov.uk>