Refactor Bigquery::Exporter
This vastly reduces the scope of the Exporter class to reflect that we
now have the IndividualExport class where most of the logic has been
moved to.

The tests for this rely quite heavily on mocking so they could be a
little fragile - I'm going to see if I can get an integration test in
the rake tasks.

Co-authored-by: Chae Cramb <chae.cramb@digital.cabinet-office.gov.uk>