Class for backfilling BigQuery export data
This class has been added to allow populating data to BigQuery that
wasn't ready at the initial set-up. It exports all the data that exists
up until the most recent export.

The motivation is for adding a new table - early_access_users - that
will contain data on individual users.

We could also use this approach to change a schema as it deletes any
existing BigQuery table.

Co-authored-by: Chae Cramb <chae.cramb@digital.cabinet-office.gov.uk>