Replace Bigquery::Resetter with rake tasks
This provides the functionality through rake tasks to delete an
individual BigQuery table or all of our export data. The resetter
previously only allowed the latter - however as we are dropping a table
(waiting_list_users) we need the ability to drop just a single table.

I considered using a class for this but it did so little that it seemed
to make sense to just do both actions directly from the rake task.