Modify bigquery:export task for earlier changes
This changes the task to use the new class that is returned from the
Exporter class.

I've also modified the test to now be an integration test as I was
uncomfortable that there was nothing that tested the integration between
the various classes involved - hence the need for BiqQuery and Smart
Survey API stubbing.