Add early_access_users to exporting data
This adds early_access_users to the records of data that is exported.
Unlike questions and answer_feedback the data in EarlyAccessUsers is
mutable so we can't export all the data otherwise, we need to try keep
BigQuery in sync with changes (yuck). Early Access Users also contains
personally identifiable information that we don't want to export.

We therefore just export a few fields that we don't expect to change and
can be used to evaluate their chat usage reasons.

To help inform a decision on whether users are external or internal we
export their domain of the email address if that domain ends in gov.uk -
this should help us identify government users - well those that used
their work address.